Exer02.py
Create a Python program that will calculate the change of the customer based on the given
user’s information. Your program should:
1. Allow the user to specify what type of fuel. The only available fuels are “DIESEL”,
“PREMIUM”, “UNLEADED”, and “KEROSENE”. In the example below, the user chooses
PREMIUM.
2. Allow the user to specify how many liters the customer bought. In the example below,
this would be 5 and 3.
3. Allow the user to specify how much did the customer pay. In the example below, the
customer pays 1000 and 100 pesos.
4. Show the amount of change based on the customer’s money. Use float in dealing with
numbers and use Philippine pesos as a unit. If the customer pays less than the expected
fuel price, the program should prompt a message.